requirements.txt
tensorflow-cpu==2.13.0
pandas==2.0.3
numpy>=1.20
Pillow==9.5.0
scipy==1.10.1
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
keras-preprocessing==1.1.2
flask==2.3.2